Fix bug and ignore files-creation-date in DataPreprocessor
The files creation date was overwriting the value which was given to the 'date' field either through unique-query or through a json-file. On top of that, the file-creation-date functionality was not very useful either, because it did not reflect the measurement date, but the analysis date, which is the one that matches the date when the data files where downloaded or copied and pasted. For these reasons, the files-creation-date is ignored as of now, and the 'date' field can now be specified via unique-query or json-file (and not be overwritten).