Add the overvoltage to the output files
The purpose of this change is the same as for commit eaf0d0a878b00581cb95515cae56c9a996d3a303, where this feature was added to summary files. Now, we are adding it to every output file. This change avoids overwriting files of measurements which are done in the same conditions but at different overvoltages.