Give dtype when setting array attributes in DarkNoiseMeas.purge_bursts()
Most of the times, it works fine without explictly setting dtype. However, for very ill-formed data where all of the spotted peaks are part of a burst (i.e. for very noisy data), frame_idx ends up being an empty list, and so, creating a numpy array out of it will create, by default, an empty array whose dtype is equal to np.float64. This makes the FrameIDX setter arise an error, because it expects a np.int64 array (even it it's an empty one).