Change name of DataPreprocessor.process_file() to get_metadata()
With the changes introduced in the last commits, the only function of this method now is to retrieve the meta-data from the input file and return it as a dictionary. Additionally, it can retrieve its creation date. That's why it is now called get_metadata() instead of process_file(). This method documentation has been updated accordingly to the last commits.