Changes to __check_well_formedness_of_subfolder DataPreprocessor method
The first change has to do with the fact that now only json and non-json files are counted. I.e. no specific extension is required for non-json files. P.e., before, .wfm files were looked for. The second change has to do with the returned objects by this method. Now, apart from an incidents/warnings list, another list with the paths to the analysable folderpaths is returned. Every folder which contains a number of json files which matches exactly the number of expected json files, is considered to be an analysable folder.