Normalize linear-fit and gaussian-fit errors according to each GainMeas
Before, the error figure of merit (FOM) was computed using the linear-fit error distribution which followed from considering the fits to all of the GainMeas objects analyzed in the current batch. Now, the set of samples that are used to compute those distributions (linear-fit error and gaussian-fit error) are the ones that follow from computing those fits for various (peaks_to_detect, bins_no) combinations, but only for the GainMeas of interest.