Partially revert changes from commit 9cac8e
In such commit, the means and standard deviations for the linear-fit and gaussian-fit error distributions were computed separatedly for each GainMeas object, using the error samples for each one of the (peaks_to_detect, bins_no) configuration. Now, we are shifting back to how things worked before that commit, were the mean and the standard deviation of each one of those error distributions was computed using the error samples from every GainMeas in the analyzed batch.