Handle RuntimeError and NotEnoughFitSamples exceptions jointly
These exceptions are now handled by trying new calls to DarkNoiseMeas.analyze() iteratively lowering the number of bins used for the amplitudes histogram. If this loop does not give a successful analysis (i.e. a call to analyze() without a RuntimeError or a NotEnoughFitSamples exception), then a new loop of calls to DarkNoiseMeas.analyze() is tried, but this time rising the number of standard deviations that are considered to delimit the gaussian fitting range. Note that, before, these exceptions were handled in parallel, i.e. the RuntimeError exception tried the std-no loop, while the NotEnoughFitSamples exception tried the bins-no loop. Now, if any of those exceptions happens, both loops happen sequentially: first the bins-no loop and, if the exception persists, then the std-no loop.